moved code from the init file out into other files
